requirements.txt
requests==2.31.0
python-dotenv==1.0.0
gtts==2.4.0
pygame==2.5.2
